refactor: extract migration logic into migrate.py
The setup path in __init__.py was carrying ~150 lines of v1→v2
migration plumbing that only runs at upgrade time. Move it to a
dedicated module so __init__.py stays focused on the hot setup path
and the migration concerns live next to each other.

- `migrate.py` exposes `migrate_mpris_unique_ids` and
  `migrate_mpris_service_mappings` (plus the private helpers/regexes).
- `async_migrate_entry` stays in `__init__.py` as a thin wrapper —
  HA's entrypoint convention preserved, no re-export tricks needed.
- Tests follow the source split: `tests/test_migrate.py` holds the
  two migration suites (16 tests, all unchanged).
